+++ src/main/java/com/github/felipeanjosunifil/projeto_mvc/model/service/ProdutoService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface ProdutoService {
     Boolean novoProduto(Produto produto) throws Exception;
     List<Produto> getProdutos();
-    Produto getProdutoPorId(int id) throws Exception;
+    Produto getProdutoPorId(Long id) throws Exception;
     List<Produto> consultarProdutos(String nome, double valorMaximo);
-    Boolean apagarProduto(int id) throws Exception;
+    Boolean apagarProduto(Long id) throws Exception;
 }
